requirements.txt
fastapi==0.115.0
httpx==0.27.2
python-dotenv==1.0.1
uvicorn==0.30.6
pytest==8.3.2
